card footers and edit mode added
