+++ src/App.css
@@ -1,0 +1,6 @@
+.cardContainer {
+  display: flex;
+  flex-wrap: wrap;
+  width: 100vw;
+  flex-direction: row;
+}
