+++ src/App.js
@@ -2,10 +2,8 @@ import "./App.css";
 import { Route, Routes } from "react-router-dom";
 import UserContextProvider from "./contexts/UserContext";
 import Homepage from "./pages/HomePage";
-import ServiceCard from "./components/Cards/ServiceCard/ServiceCard";
 import { createTheme, ThemeProvider, styled } from "@mui/material/styles";
 import { green, purple, yellow, grey } from "@mui/material/colors";
-import PhotographerCard from "./components/Cards/PhotographerCard";
 import "@fontsource/raleway"; // Defaults to weight 400.
 
 const theme = createTheme({
