+++ src/components/Cards/PhotographerCard/PhotographerCard.css
@@ -1,5 +1,3 @@
 .cardCover {
-  border-bottom-left-radius: 0.75rem;
-  border-bottom-right-radius: 0.75rem;
   box-shadow: rgba(0, 0, 0, 0.3) 0px 3px 7px -3px;
 }
