+++ src/components/Cards/PhotographerCard/PhotographerCard.jsx
@@ -1,30 +1,14 @@
 import * as React from "react";
-import { styled } from "@mui/material/styles";
 import "./PhotographerCard.css";
 import Card from "@mui/material/Card";
 import CardHeader from "@mui/material/CardHeader";
 import CardMedia from "@mui/material/CardMedia";
-import CardContent from "@mui/material/CardContent";
-import CardActions from "@mui/material/CardActions";
-import Collapse from "@mui/material/Collapse";
 import Avatar from "@mui/material/Avatar";
 import IconButton from "@mui/material/IconButton";
 import Typography from "@mui/material/Typography";
-import ExpandMoreIcon from "@mui/icons-material/ExpandMore";
 import StarIcon from "@mui/icons-material/Star";
-import FavoriteIcon from "@mui/icons-material/Favorite";
 import Skeleton from "@mui/material/Skeleton";
-
-const ExpandMore = styled((props) => {
-  const { expand, ...other } = props;
-  return <IconButton {...other} />;
-})(({ theme, expand }) => ({
-  transform: !expand ? "rotate(0deg)" : "rotate(180deg)",
-  marginLeft: "auto",
-  transition: theme.transitions.create("transform", {
-    duration: theme.transitions.duration.shortest,
-  }),
-}));
+import PhotographerCardFooter from "../PhotographerCardFooter";
 
 export default function PhotographerCard({
   name,
@@ -33,15 +17,13 @@ export default function PhotographerCard({
   profilePic,
   location,
   isLoaded,
+  withFooter,
 }) {
-  const [expanded, setExpanded] = React.useState(false);
-
-  const handleExpandClick = () => {
-    setExpanded(!expanded);
-  };
-
   return (
-    <Card sx={{ maxWidth: 250, borderRadius: 2 }} elevation={4}>
+    <Card
+      sx={{ minWidth: 250, maxWidth: 250, borderRadius: 2, margin: 2 }}
+      elevation={4}
+    >
       <CardHeader
         className="cardHeader"
         font
@@ -108,43 +90,7 @@ export default function PhotographerCard({
       ) : (
         <Skeleton variant="rectangular" height={194} align="center" />
       )}
-
-      <CardActions disableSpacing sx={{ p: 0 }}>
-        <IconButton aria-label="add to favorites" sx={{ p: 0.25 }}>
-          <FavoriteIcon sx={{ color: "tertiary.main" }} />
-        </IconButton>
-        <ExpandMore
-          sx={{ p: 0 }}
-          expand={expanded}
-          onClick={handleExpandClick}
-          aria-expanded={expanded}
-          aria-label="show more"
-        >
-          <ExpandMoreIcon />
-        </ExpandMore>
-      </CardActions>
-      <Collapse in={expanded} timeout="auto" unmountOnExit>
-        <CardContent>
-          <CardMedia
-            component="img"
-            height="194"
-            image="https://i.insider.com/61a7a6e10ed48c0019e537e8?width=1000&format=jpeg&auto=webp" //image={img} Esta linea es para ya pasar la info
-            alt="Paella dish"
-          />
-          <CardMedia
-            component="img"
-            height="194"
-            image="https://i.insider.com/61a7a6e10ed48c0019e537e8?width=1000&format=jpeg&auto=webp" //image={img} Esta linea es para ya pasar la info
-            alt="Paella dish"
-          />
-          <CardMedia
-            component="img"
-            height="194"
-            image="https://i.insider.com/61a7a6e10ed48c0019e537e8?width=1000&format=jpeg&auto=webp" //image={img} Esta linea es para ya pasar la info
-            alt="Paella dish"
-          />
-        </CardContent>
-      </Collapse>
+      {withFooter ? <PhotographerCardFooter /> : <></>}
     </Card>
   );
 }
