+++ src/components/Cards/PhotographerCardFooter/index.jsx
@@ -1,0 +1,1 @@
+export { default } from "./PhotographerCardFooter";
