+++ src/components/Cards/ServiceCard/ServiceCard.jsx
@@ -6,11 +6,22 @@ import CardMedia from "@mui/material/CardMedia";
 import Typography from "@mui/material/Typography";
 import { CardActionArea } from "@mui/material";
 import Skeleton from "@mui/material/Skeleton";
+import ServiceCardFooter from "../ServiceCardFooter";
 
-export default function ServiceCard({ service, img, isLoaded }) {
+export default function ServiceCard({
+  service,
+  img,
+  isLoaded,
+  withFooter,
+  minPrice,
+  editMode,
+}) {
   return (
     <>
-      <Card sx={{ maxWidth: 250, borderRadius: 2 }} elevation={4}>
+      <Card
+        sx={{ maxWidth: 250, borderRadius: 2, minWidth: 202, margin: 2 }}
+        elevation={4}
+      >
         <CardActionArea sx={{ maxWidth: 300, padding: ".5rem" }}>
           {isLoaded ? (
             <CardMedia
@@ -38,6 +49,15 @@ export default function ServiceCard({ service, img, isLoaded }) {
             )}
           </CardContent>
         </CardActionArea>
+        {withFooter ? (
+          <ServiceCardFooter
+            editMode={editMode}
+            isLoaded={isLoaded}
+            minPrice={minPrice}
+          />
+        ) : (
+          <></>
+        )}
       </Card>
     </>
   );
